% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndata.R
\docType{package}
\name{bndata}
\alias{bndata}
\alias{bndata-package}
\title{bndata: Curated Datasets for Bayesian Networks}
\description{
bndata is a centralized repository of publicly available datasets
used to construct Bayesian networks. Mainly intended as a centralized
resource for conducting Bayesian network research and for reproducing
existing results.
}

